unify dozent and tutor component
