+++ web-gui/src/app/components/courses/grant-docent-snipp/grant-docent-snipp.component.html
@@ -1,0 +1,26 @@
+<mat-chip-list>
+  <mat-chip
+    *ngFor="let docent of course.course_docent"
+    color="accent"
+    selected="true"
+    removable="true"
+    (removed)="removeDocent(course.course_id, docent.user_id)">
+    {{docent.prename}} {{docent.surname}}
+    <mat-icon matChipRemove>cancel</mat-icon>
+  </mat-chip>
+
+  <mat-form-field *ngIf="showInputForDocent && course.course_id === docentInputCourseID">
+    <input type="text" matInput [formControl]="docentFormControl" [matAutocomplete]="auto"
+           placeholder="User auswählen" (keyup)="addDocent(course.course_id, $event.key)">
+  </mat-form-field>
+
+  <mat-autocomplete #auto="matAutocomplete" [displayWith]="displayFn">
+    <mat-option *ngFor="let option of filteredOptions | async" [value]="option">
+      {{option.surname}} {{option.prename}}
+    </mat-option>
+  </mat-autocomplete>
+
+  <button mat-icon-button color="accent"  (click)="showDocentInput(course.course_id)">
+    <mat-icon>add_circle</mat-icon>
+  </button>
+</mat-chip-list>
