+++ web-gui/src/app/components/courses/grant-docent-snipp/grant-docent-snipp.component.ts
@@ -1,0 +1,102 @@
+import {Component, EventEmitter, Input, OnInit, Output, ViewChild} from "@angular/core";
+import {MatSnackBar, MatSort} from "@angular/material";
+import {User} from "../../../interfaces/HttpInterfaces";
+import {flatMap, map, startWith} from "rxjs/operators";
+import {DatabaseService} from "../../../service/database.service";
+import {UserService} from "../../../service/user.service";
+import {FormControl} from "@angular/forms";
+import {Observable} from "rxjs";
+
+@Component({
+  selector: 'app-grant-docent-snipp',
+  templateUrl: './grant-docent-snipp.component.html',
+  styleUrls: ['./grant-docent-snipp.component.scss']
+})
+export class GrantDocentSnippComponent implements OnInit {
+  @ViewChild(MatSort) sort: MatSort;
+  @Input() course;
+  @Output() loadAllCourses: EventEmitter<void>;
+
+  docentFormControl = new FormControl();
+  filteredOptions: Observable<User[]>;
+
+  //dataSourceCourses = new MatTableDataSource<GeneralCourseInformation>();
+  dataSourceUsers : User[];
+  showInputForDocent: boolean;
+  docentInputCourseID: number;
+
+  constructor(private db: DatabaseService, private user: UserService, private snackBar: MatSnackBar) {
+    this.loadAllCourses = new EventEmitter<void>()
+  }
+
+  ngOnInit() {
+    this.db.getAllUsers().pipe(
+      flatMap(users => {
+        this.dataSourceUsers = users;
+        this.filteredOptions = this.docentFormControl.valueChanges
+          .pipe(
+            startWith<string | User>(''),
+            map(value => typeof value === 'string' ? value : value.prename.concat(value.surname)),
+            map(name => name ? this._filterDocentInput(name) : this.dataSourceUsers.slice())
+          );
+        return this.filteredOptions;
+      })
+    ).subscribe();
+  }
+
+  private _filterDocentInput(value: string): User[] {
+    const filterValue = value.toLowerCase().replace(' ', '');
+
+    return this.dataSourceUsers.filter(option => {
+      return option.prename.toLowerCase().indexOf(filterValue) === 0
+        || option.surname.toLowerCase().indexOf(filterValue) === 0
+        || option.surname.toLowerCase().concat(option.prename.toLowerCase()).indexOf(filterValue) === 0
+        || option.prename.toLowerCase().concat(option.surname.toLowerCase()).indexOf(filterValue) === 0;
+    });
+  }
+
+  /**
+   * Add docent to course
+   * @param courseID Course, the docent will be added
+   * @param key Keyboard press key 'ENTER'
+   */
+  addDocent(courseID: number, key: string) {
+    if (key === 'Enter') {
+      const selectedUser: User = this.docentFormControl.value;
+      this.docentFormControl.setValue('');
+      this.showInputForDocent = false;
+
+      this.db.addDocentToCourse(courseID, selectedUser.user_id).subscribe(res => {
+        this.loadAllCourses.emit();
+      })
+    }
+  }
+
+  /**
+   * Remove docent from course
+   * @param courseID Course, docent will be removed
+   * @param userID The docent id
+   */
+  removeDocent(courseID: number, userID: number) {
+    this.db.removeDocentFromCourse(courseID, userID).subscribe(courses => {
+      this.loadAllCourses.emit();
+    });
+  }
+
+  /**
+   * Show only for course with right id the input
+   * @param courseID The course input should be shown
+   */
+  showDocentInput(courseID: number) {
+    this.docentInputCourseID = courseID;
+    this.showInputForDocent = true;
+  }
+
+  /**
+   * Shows prename and surname of users in autocomplete
+   * @param user The user to show
+   */
+  displayFn(user?: User): string | undefined {
+    return user ? user.prename + ' ' + user.surname : undefined;
+  }
+}
