+++ web-gui/src/app/components/courses/grant-docent/grant-docent.component.html
@@ -1,63 +1,26 @@
 <div class="container">
-
-
   <mat-form-field appearance="outline">
     <mat-label>Module filtern</mat-label>
     <input matInput (keyup)="filterCourses($event.target.value)">
     <mat-icon matSuffix>filter_list</mat-icon>
   </mat-form-field>
   <table mat-table [dataSource]="dataSourceCourses" matSort>
-
     <ng-container matColumnDef="course_name">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>
         <h2>Module</h2>
       </th>
       <td mat-cell *matCellDef="let course">
         <h3>{{course.course_name}}</h3>
-
-
-        <app-grand-tutor-snipp (loadAllCourses)="loadAllCoursesTutor($event)"  [course]="course"></app-grand-tutor-snipp>
-
+        <mat-divider></mat-divider>
+        <h4>Tutoren</h4>
+        <app-grant-tutor-snipp (loadAllCourses)="loadAllCourses($event)"  [course]="course"></app-grant-tutor-snipp>
         <mat-divider></mat-divider>
         <h4>Dozenten</h4>
-
-        <mat-chip-list>
-          <mat-chip
-            *ngFor="let docent of course.course_docent"
-            color="accent"
-            selected="true"
-            removable="true"
-            (removed)="removeDocent(course.course_id, docent.user_id)">
-            {{docent.prename}} {{docent.surname}}
-            <mat-icon matChipRemove>cancel</mat-icon>
-          </mat-chip>
-
-          <mat-form-field *ngIf="showInputForDocent && course.course_id === docentInputCourseID">
-            <input type="text" matInput [formControl]="docentFormControl" [matAutocomplete]="auto"
-                   placeholder="User auswählen" (keyup)="addDocent(course.course_id, $event.key)">
-          </mat-form-field>
-
-          <mat-autocomplete #auto="matAutocomplete" [displayWith]="displayFn">
-            <mat-option *ngFor="let option of filteredOptions | async" [value]="option">
-              {{option.surname}} {{option.prename}}
-            </mat-option>
-          </mat-autocomplete>
-
-
-          <button mat-icon-button color="accent"  (click)="showDocentInput(course.course_id)">
-            <mat-icon>add_circle</mat-icon>
-          </button>
-        </mat-chip-list>
-
-
+        <app-grant-docent-snipp (loadAllCourses)="loadAllCourses($event)"  [course]="course"></app-grant-docent-snipp>
       </td>
     </ng-container>
 
-
     <tr mat-header-row *matHeaderRowDef="columns"></tr>
     <tr mat-row *matRowDef="let row; columns: columns;"></tr>
-
   </table>
-
-
 </div>
